+++ eslint.config.mjs
@@ -191,7 +191,7 @@ export default [
             "playwright/no-force-option": "error",
             "playwright/no-raw-locators": "error",
             "playwright/expect-expect": [
-                "warn",
+                "error",
                 {
                     "assertFunctionPatterns": ["^expect.*", "^assert.*"]
                 }