+++ src/Glpi/Api/HL/Controller/AdministrationController.php
@@ -1349,16 +1349,14 @@ EOT,
      * @param string|null $picture_path The path to the picture from the user's "picture" field.
      * @return Response A response with the picture as binary content (or the placeholder user picture if the user has no picture).
      */
-    private function getUserPictureResponse(string $username, ?string $picture_path): Response
+    private function getUserPictureResponse(Request $request, string $username, ?string $picture_path): Response
     {
         if ($picture_path !== null) {
             $picture_path = GLPI_PICTURE_DIR . '/' . $picture_path;
         } else {
-            $picture_path = 'public/pics/picture.png';
+            $picture_path = GLPI_ROOT . '/public/pics/picture.png';
         }
-        $symfony_response = Toolbox::getFileAsResponse($picture_path, $username);
-
-        return new Response($symfony_response->getStatusCode(), $symfony_response->headers->all(), $symfony_response->getContent());
+        return $this->downloadFile($request, $picture_path, $username);
     }
 
     #[Route(path: '/User/Me/Picture', methods: ['GET'], scopes: ['user'])]
@@ -1377,7 +1375,7 @@ EOT,
             ],
         ]);
         $data = $it->current();
-        return $this->getUserPictureResponse($data['name'], $data['picture']);
+        return $this->getUserPictureResponse($request, $data['name'], $data['picture']);
     }
 
     #[Route(path: '/User', methods: ['POST'])]
@@ -1420,7 +1418,7 @@ EOT,
             ],
         ]);
         $data = $it->current();
-        return $this->getUserPictureResponse($data['name'], $data['picture']);
+        return $this->getUserPictureResponse($request, $data['name'], $data['picture']);
     }
 
     #[Route(path: '/User/username/{username}/Picture', methods: ['GET'], requirements: ['username' => '[a-zA-Z0-9_]+'])]
@@ -1439,7 +1437,7 @@ EOT,
             ],
         ]);
         $data = $it->current();
-        return $this->getUserPictureResponse($data['name'], $data['picture']);
+        return $this->getUserPictureResponse($request, $data['name'], $data['picture']);
     }
 
     #[Route(path: '/User/{id}', methods: ['PATCH'], requirements: ['id' => '\d+'])]