+++ tests/e2e/specs/Form/DestinationConfig/actors.spec.ts
@@ -190,6 +190,7 @@ for (const actor_type of actor_types) {
             }
         });
 
+        /* eslint-disable-next-line playwright/expect-expect -- expects are done in doAssertDropdownValueIsNotAvailable */
         test('Cannot select unauthorized actors with "Specific actors"', async () => {
             await form_page.doOpenDestinationAccordionItem('Actors');